Label scan store is no longer a kernel extension
As there is only one label scan store implementation
there is no need to keep it around as a kernel extension.

Remove
 - LabelScanStoreProvider
 - NativeLabelScanStoreExtension
 - LabelIndex enum in settings

Instead of being loaded as a kernel extension label scan store
is constructed in RecordStorageEngine.

DynamicIndexStoreView that decorate NeoStoreIndexStoreView
with the possibility to utilize LabelScanStore for some operations
used to do this by extending NeoStoreIndexStoreView. This
created a circular dependency that went like

LabelScanStore -> FullLabelStream -> IndexStoreView -> LabelScanStore

This dependency is broken apart by letting DynamicIndexStoreView
delegate to NeoStoreIndexStoreView instead of extending it.
This makes it possible for NativeLabelScanStore to use
NeoStoreIndexStoreView instead of DynamicIndexStoreView.