Limit DynamicArray chunk size to less than Integer.MAX_VALUE
Even though arrays are indexes by ints, they don't actually have the full 31
bit range available for their length. It's JVM implementation specific, but the
maximum array size is somewhere slightly less than Integer.MAX_VALUE.
For this reason we specify the maximum chunk size to be Integer.MAX_VALUE minus
Short.MAX_VALUE. This is more than enough room for at least HotSpot, as far as
the author is aware.

No test because of how implementation specific it is, and how trivial the
change is.