PBI: Page cache backed NumberArrayFactory, first draft
This adds a new NumberArrayFactory that builds number arrays that use the page
cache as backing memory. This is only used a the last fall-back for the
ChunkedNumberArrayFactory for allocating chunks, so it will be used as little
as possible. However, it will allow the importer to import graphs on systems
where for instance the IdMapper cannot fit in the available memory of the
system.

There is one outstanding issue in this commit, and that is that the
NumberArrays can be exposed to concurrent access from multiple threads. The
PageCache backed NumberArray implementations are currently single-threaded,
becuase they reuse the same pre-allocated PageCursors over and over again. If
the array is accessed concurrently, then this can cause a cursor to move while
another thread is reading from the pinned page.