Fix a bug where the importer could produce an inconsistent store
As an import stage approaches completion, the steps starts to shut down
following the reception of the last batch. The last batches have to be
processed and sent downstream before a step can be said to be done.

Steps that spawn threads to process batches in parallel now rely on a WorkSync
to ensure that processed batches are sent downstream in order. The WorkSync
sorts the batches before sending them downstream, since the processing may
complete out of order. A "null-batch" is sent as the last one, to signal to the
WorkSycn that no more batches are coming. This batch is queued up when a step
begins its transition to done.

The problem was that this was done _after_ the AbstractStep had told the
downstream step, that it had reached the end of the upstream, and thus should
expect no more batches. Thus, the shutting down of the downstream step was
racing with the processing of the last set of batches from a given parallel
step.

This meant that arbitrary batches could be processed only partially, in any
stage that involved a ForkedProcessorStep. This could even afflict batches that
are otherwise in the "middle" of the dataset, due to multi-pass processing.
Likewise, the shared-nothing design of the counts store stage meant that very
large fractions of the total counts could be missing.

The fix is to ensure that a step doesn't tell a downstream step to shut down,
before that step considers itself to be fully done processing data.