Appends tree information on exceptions leaving surface
There are many places where exceptions can be thrown from inside internal
code related to GBPTree and its internals. However the surface of GBPTree
is very small and so the approach is to not pass in tree information into
all internal classes and static methods, but instead decorate exceptions
from the small number of places where these exceptions can leave the
surface. This approach is much less intrusive than the alternative.

This is how it was already, but not the entire surface was covered.
Now also SeekCursor decorates outgoing exceptions.