Importer option to disallow cache allocation on heap
Normally, allocating data arrays for caching in the importer tries off-heap,
falls back to on-heap and then on to additional ways of allocating.
This commit introduces a config option in the import tool to disallow
the on-heap alternative. This to have better control over heap usage
for normal working memory.

Configuration option is e.g. `--cache-on-heap=false`