IdMapper no longer keeps collision values in heap
The input ids from the node import (:ID) are encoded into a more
memory-efficient representation for mapping :ID --> node record id.
Some input ids happens to be duplicates and even some encoded ids will
accidentally be the same. For collisions within the same input group
the actual input ids needs to be compared to see whether or not they
are actual or accidental collisions.

Previously this was done by going through the input one more time,
from start to end and collecting the collision-marked ids into memory
for duplicate detection as well as during the input id --> node id
lookups later. In the event of lots of collisions, which btw increases
with number of imported nodes (the accidental collisions), this may
require a lot of memory. The second pass of the input is also suboptimal
performance-wise.

This is changed such that input ids are stored into a temporary
property store next to the imported db where there's a simple 1-to-1
mapping between node and input id property record.
This has a slight cost to the node import and relationship import stages,
but on the flip side speeds up IdMapper preparation stage and cuts down
on heap memory usage in scenarios where there are lots of input id collisions.