Empty stream instead of exception for non-existing directory
Return empty stream instead of throwing an exception in case if the
directory does not exist. This will gonna allow the creation of empty
streams on top of non-existing directories and will allow for example
usage of stream find/match methods instead of catching the exception to
determine the correctness of situation in user code.