Introduces CommandCreationContext
Which has context around committing and command creation
CommandCreationContext owner is StorageStatement. If may feel slightly funky,
but the reason is that it now also (instead of StoreReadLayer) is responsible for
reserving node/relationship ids, which can now use this context too. Also during commit
the StorageEngine#createCommands already has access to the StorageStatement.

Some commit structures, like PropertyCreator and friends were previously global
for a RecordStorageEngine instance, but is now local to its StorageStatement instead.
This way each statement can instantiate these structures with its own
CommitCreationContext and so inject the id generators that it has.