Renamed combined index provider to fusion
while keeping it generic, even more generic than before by passing in
a selector between one and the other. The idea is that specific combinations
of providers will live in specific kernel extensions. To have the fusion
provider generic is nice from a single responsibility perspective where
it has no idea about what the providers are, it just combines the two.