Importer uses correct page cache in caching arrays
Previously the normal use case of the ParallelBatchImporter would result in
a null PageCache given to PagedCachedNumberArrayFactory and would result in
NPE when allocation got to the point of allocating into it.

Now the page cache use by BatchingNeoStores, be it the external page cache
given to the PBI, or otherwise the one constructed by itself will be passed
to that factory.