TransactionBoundQueryContext no longer creates new statements
Instead always uses the statement passed in. This was a problem because
isDeleted methods created a new statement, but didn't close it in the end.
This was both unnecessary and it created problems due to causing statements
to be open longer than they should, causing e.g. index reader caching
to hold on to cache readers too long.