Use WordUtils from commons-text
Arguments depended on a buggy implementation of WordUtils.wrap() where
leading whitespace on new lines was not stripped despite the docs saying it should.
In commons-lang3 3.6 WordUtils is deprecated and moved to another module
commons-text. This commit introduces that module and solves the leading-space
problem for lines that do not need to be wrapped by not calling WordUtils.wrap
for them. Lines that need wrapping and start with leading whitespace
(indentation) will still have it stripped going forward.