Add IncrementalXXH64 as an option for producing index entry ids
This can help alleviate situations where many transactions with a large number of updates via merge, would end up with false deadlocks due to collisions on their index entry hashes.
We cannot use the new hash function by default because that would be an HA protocol change, but we can hide it behind a feature toggle.