Check only first property for uniqueness during uniqueness constraint verification
Before we were checking all the fields of index for uniqueness
when we were about to create uniqueness constraint. That was fine for
one property index but n case of composite index of N properties
we would actually do verification N times, while checking combinations
for first property only is more then enough.