Replication throttling
Puts a limit on how much data can be in the replication phase concurrently
and at the same time improves the messaging stack so that the send calls
have the option to block. This allows better control over resends which
previously could be attempted even before the queued messages had
hit the network. Because the channel is shared with time critical
messages it also allows a stable behaviour under situations where alot
of concurrently replicating activity is ongoing, which with a limit
in place will allow the time critical messages to mesh better into the flow.

This is a medium-term solution which will be improved.