LogTailScanner updated
LogTailScanner is now invoked once during startup to collect information about the tail of the transaction logs, and this result it reused by all component interested in such information.

If any log entries with unsupported version is found, a better error message is thrown to avoid confusion as to what went wrong.

We now also properly respect dbms.allow_upgrade when the transaction log format is changed.