First stab at adding Iterator Stability isolation level
There are a lot of complications in this area due to having both the lock
clients and the LockService, which is used for short-lived latches.

The reason we have the lock service, I think, is to avoid propagating
short-lived locks in HA, and technically they did not need any deadlock
resolution... until now.

Long term, the LockService should be removed entirely, so we only use the lock
clients. We should then add API to the lock clients such that we can say up
front whether a lock is meant to be short-lived or long-lived, where the
SlaveLocksClient, for instance, would only propagate the long-lived locks
throughout an HA cluster.

Other components that currently rely on the LockService, would then instead
need to be able to acquire their own Locks.Client instances, and participate in
dead lock resolution like everyone else, which should not be a problem for
them, since they only take one lock at a time. That is, assuming Forseti
doesn't throw false positive dead locks at them. They will have to grow retry
loops for dealing with that.