Specialize SequenceValue equals and compareTo
Some SequenceValue implementations have list like behaviour, and do thus
not support constant speed random access. In these cases it will be much
faster to iterate over the sequence using an Iterator.

We introduce a way to discover the preferred method of iteration from any
sequence value, and specialize the SequenceValue equals and compareTo so
that they are always O(n+m), where n and m are the sizes of the
sequences.