Fix variable scope bug with aggregation, foreach and sorting
For certain queries with aggregation expressions in combination with
child scopes (FOREACH) and ORDER BY there was a problem with variable
scope resulting from the interaction of the AST rewriters
normalizeWithClauses, expandStar and isolateAggregation with semantic
analysis.
The expandStar rewriter would rewrite PragmaWithout clauses to explicit
With clauses using variable positions that would later be outdated by
the second semantic analysis that follows rewriting, which would break
the correctness of that semantic analysis.