Make the free id keeper work as a stack in aggressive mode
The free id keeper used to append at the end and eat from the
beginning. This leads to growth of the id file. This behaviour
will stay for the regular mode. It is mitigated in the regular
mode by compacting the gap during close.

For the aggressive mode however the keeper will now append to
the end and aggressively eat as well as truncate from the end
as batches get written to and read from disk.