Work-around for broken legacy db in migration test
LegacyIndexesUpgradeTest uses a zipped legacy db to test migration from.
That db has some very weird meta-data records which makes e.g. log version 1
skipped, so db has version 0 and will go to version 2 right away. Also
There are transactions in it but the lastCommittedTxId field(s) says
that it's 1 (i.e. no transactions in it at all).This causes the db to
break on the second start up, which will cause recovery and will silently
ignore the missing log version 1 for some reason and happen to complete.

The reverse log reader is more strict and so fails. Also that recovery
that happens actually performs recovery on old log entries, which is not
allowed whatsoever. The work-around here is to start the db only once,
which works and where it can verify that manual index migration worked.

This db should die in flames and be recreated. Unfortunately there's no
code embedded in the test to create it. The db is legacy-index-db.zip