Fix statement acquisition in token accessing built-in procedures
The built-in procedures using `TokenAccess` (`db.labels()`,
`db.relationshipTypes()`, `db.propertyKeys()`) would close their
acquired `KernelStatement` prematurely before returning an iterator
stream was using the statement and that would close it when exhausted.
This caused a problem when a multiple of these procedure calls were
made in the same query. Changed this so that ownership of the acquired
statement reference is transferred to the iterator (which is consistent
with the core API call usage of `TokenAccess`).

Also changed the `TokenIterator.close()` to be idempotent, since
exhausting the procedure stream also causes close to be invoked.