Add logic for handling side effects according to standard
Previously, side effects in the openCypher TCK were validated towards
the QueryStatistics interface, which is part of the Neo4j Java API. It
has since been proposed that these elements should be defined in a
Cypher-only way, and validated through fixed Cypher queries. This commit
adds functionality for dealing with side effects in such a way, but does
not hook these into the TCK execution yet.