Changes schema index directory structure for fusion index
The lucene schema index provider puts its indexes in:

  `<db>/schema/index/lucene/<indexId>/...`

this commit doesn't change this, but the fusion index provider,
i.e. the combination of lucene+native gets a new structure driven by the need
to one day be able to do efficient background migration of schema indexes
and for convenience in the wild. The new structure looks like this:

  `<db>/schema/index/<providerKey>/<providerVersion>/<indexId>/...`

Including the sub-provider directory and specifically for this fusion index (for indexId 5):

  `<db>/schema/index/lucene_native-1.0/5/lucene-1.0/...`
  `<db>/schema/index/lucene_native-1.0/5/native-1.0...`

This has two benefits:

- all index data related to one index lives under one directory
- indexes on disk are also versioned, just like providers are

Since the fusion index also uses a lucene schema index provider internally,
this structure has been made pluggable and given to a SchemaIndexProvider
at construction time. See IndexDirectoryStructure. This class has all knowledge
about the base schema root directory and various directory structure implementations
collected in one place.

NOTE This change does not require any migration, except for databases recently
created, where the fusion index is enabled by default. Although no releases
has had the fusion index in them so that is not a problem.