Virtual node hack to support db.schema procedure
The db.schema procedure has been implemented using a hack where virtual
nodes and relationships were created with negative ids. This was done
in order to display the results as a graph in the browser, and has been
working well as long as we have been using the Core API types in Cypher.

Upon introducing the Value types, this did not work anymore, as the hacked
virtual node information was lost during conversions.

This commit introduces a hack to pass on the hacked Core API Node/Relationship
implementations to the surface if they have negative id numbers. It should/can
be removed when we introduce support for full stack virtual nodes.

We are not proud...