Improve log assertion in BoltFailuresIT
Test `boltServerLogsRealErrorWhenDriverIsClosedWithRunningTransactions`
was flaky because it did not properly wait for error message.
It asserts that particular error message appears in logs after a driver
transaction failure. This failure can be observed earlier then error
is logged.

This commit makes test wait up to 1 minute for the message to appear.