Add VerboseTimeout rule in SessionResetIT
This test is flaky only on IBM JDK8 and it's unclear what thread can
be waiting for 5 minutes. This commit lowers waiting time to 1 minute
and adds verbose timeout so we can see full thread dump on failure.
Might be that query execution takes longer and we need to make it
load smaller CSV.