Using a predicate to allow more precise assertions for plans in CypherComparisonSupport
Instead of comparing similarity of plans, we expect test writers
to provide an explicit predicate or assertion that will be evaluated
on the plans. You still have to specify if there is a subset of the
configuration where you expect the predicate/assertion to fail.

This introduces the use of the aforementioned predicate in
EagerizationAcceptanceTest and UniqueIndexAcceptanceTest. These are
the ones that previously have been using succeedWithAndAssertPlansAreSimilar.