Support OpenSSL
OpenSSL still has to be linked to using the guide here:
  http://netty.io/wiki/forked-tomcat-native.html

which has several alternatives, including using the openssl
version installed on the system or statically linked deps.

The test uses a test-only dependency on boring ssl.

It is enough to put the desired version into plugins and
change the SSL provider in config to get it functional.