Use online index sampling when populating from Label index
When label index was implemented on top of Lucene, the data during scan of label index
could be obsolete resulting in multiple updates for same node during population.
Indexes themselves could handle this, but samplers could not. Therefore sampling was done
as a separate step after population was done.
Since Label index is implemented on top of GBPTree now, this is not longer necessary.