Fixes broken concurrency in SchemaCache
Previously there were only some data structures (maps/sets) that used implementations
which could handle concurrent access. Seemingly that was added ad-hoc.
Originally the state in this class was meant to not care about concurrency
and instead rely on external synchronization. However this was never fully
the case since there could be unguarded reads happening concurrently with updates
to these non-concurrency-aware data structures, which means that some existing
data could be observed as missing when other unrelated data items were removed
or added. This was indeed proven by ConcurrentCreateDropIndexIT in an upcoming
commit.