Ignore election performance IT
Was already ignored for the rapid conditions and are now ignored
for the normal conditions as well, since running builds in parallel
has a significant effect on the outcome.

The intention is to move this to a proper benchmarking environment
eventually. The tests are left in place so that one can run them
manually until then.

The test still exists unignored on older branches where parallel
testing is not utilized.