Removes optimization to not checkpoint GBPTree on no changes
This solves an issue which would be observed as a tree looking like non-clean
even after a clean db shutdown. The main reason for this is that skipping
the checkpoint would not await the background clean job and so closing the tree
could potentially be done before the clean job was done. This would prevent
the tree from being marked as clean on close.

Read-only tools like consistency checker asserts that no recovery is required
before starting and so would fail on a db containing such a tree.