Make the ForsetiFalseDeadlockTest more robust
Previously the test would fail on any false deadlock that popped out.
Now, when a false deadlock is detected, the test instead does a more thorough
check, to make sure that the false deadlock failure rate is 2% or less.

The 2% was chosen because that roughly matches the flaky failure rate we have
been observing on our build systems so far.