Removing NewPlannerTestSupport from SemanticAcceptanceTests
The original plan was to move those to CypherComparisonSupport.
We decided against that in the end. CREATE, MATCH, MERGE, and DELETE
queries are tested in separate Acceptance tests (with
CypherComparisonSupport). Here, the actual tests assert that
sequences of queries leave the database in a sane state.
CypherComparisonSupport cannot execute sequences of statements
in the same transaction. Therefore having this tested on the current
branch's versions seems good enough.