Fixes issue including seed in RandomRule
Tests that fails and also has RandomRule will have their error message
decorated with the random seed used in the test. In scenarios where
there were failure @After method or other rules failing in their "after"
code would make it so that the seed would prevent this seed from
being visible in the final printer error from the test.

This fixes that by specifically catering for this scenario, where junit
throws a specific multi-failure exception. The RandomRule will see this
and decorate all of those failures so that it becomes visible again.