Support pre-voting in Raft
Pre-voting adds robustness to the leadership elections. This is a breaking
change and must be explicitly enabled.

A Follower will only initiate an election if it has received a quorum
of positive pre-votes from other servers.

A positive pre-vote will only be returned from a server if that server
is itself asking for pre-votes. Therefore an election will be triggered
not by one server timing out, but by a majority of servers timing out.

This commit introduces a visitor pattern for RaftMessages used by
RaftMessageHandlers and RaftMessageEncoder.