Fix deadlock between index close and index add
`this` is instance of `LuceneDataSource`

T1 want to delete index 1
T2 want to add to index 1

T1 `closeIndex` synchronized on `this`
T2 `getIndexSearcher` synchronize on searcher for index 1
T1 BLOCKED on `searcher.dispose` on searcher for index 1,
   which internally synchronize on searcher. Held by T2.
T2 BLOCKED on `syncGetIndexSearcher`, synchronized on `this`.
   Held by T1.
DEADLOCK

Instead of synchronizing on found searcher, T2 synchronize
on `this` avoiding grabbing locks in reverse order.