Fixes problem with nondeterministic planning of optional matches
When multiple optional matches exist, and they do not have dependencies
between each other, the planner is free to solve them in any order. This
unfortunately causes a problem where the planner thinks it has failed to
solve the original query. This change changes the IR before planning so
we have a deterministic order to solve optional matches in.