Prevent ExplicitIndexProxy from closing statment multiple times
When using getSingle in ExplicitIndexWrapHits statement would be
closed twice because of Iterators.singleOrNull closing provided
iterator if it is a resource, which ExplicitIndexWrapHits is.

We now keep track of close state to prevent it being closed
multiple times and thereby closing statement for others that
might be using it.