Deprecate the Service.Implementation annotation
We used to have an annotation processor that would find these annotated classes and write META-INF/services files for us.
Due to problems with getting that to work with our day-to-day development tools, the annotation processor was removed.
However, this annotation remained.
Unfortunately, it is public API, so we cannot just remove it. Hence it has been deprecated.