Make the CoreReplicationIT.pageFaults... test not flaky
The problem was that the test was assuming that committing a transaction
to a causal-cluster leader would not return until the transaction was
*applied* to a cluster majority.
That is, however, not how this works. The transaction commit returns as
soon as the transaction has been replicated to the raft log of
a majority, which is the step before the transaction is applied.
This way, the test ended up racing with the transaction applier.
To solve this we simply turn the immediate assertion into an eventual
assertion.