Simplify slot allocation
Makes allocating a slot for a variable an idempotent operation,
so that individual steps in the slot allocation pass does not need
to worry about if a slot has already been allocated for a variable
within the same pipepline by a previous step.

The slot may have slightly different nullability and specific type as
long as it is compatible.
- If nullability differs, the nullability is conservatively increased.
- If types differs (but are compatible) the least specific type
is chosen.

If the slots are incompatible a compile-time exception is still thrown.