Avoid migrating "Nothing" during database upgrade.
Previously in case if NOT_PARTICIPATING participant in anyway was added
into a list of database upgrade participants it took part in database
migration and was reporting its progress as "Migrate Nothing" thingy.
This PR changes that and simply ignores NOT_PARTICIPATING participants
that added into StoreUpgrader during its constructions.