Make expression slot allocation handle two sourced plans
ValueHashJoin has one expression that is evaluated on the lhs
and one that is evaluated on the rhs. We now do two passes of
allocate expression on logical plans with two sources, one for
lhs and one for rhs (and block traversal of the expression on the side
we are not currently looking at).