Make Variable an extendable class
For the SlottedRewriter to be able to rewrite variables in expressions
where they are declared as Variable, the corresponding runtime
variable expressions (e.g. NodeFromSlot) needs to extend Variable.
To do this we change Variable from a case class to a class, and
implement the methods that case class gave us to allow us to still
treat it as a case class (when matching, rewriting etc.).