Fix ScopeExpression case with conflicting slot names
In some uniqueness predicates for variable length patterns
with a relationship iterator, the name and type of the
variable within the scoped expression for the predicate can
conflict with an outer variable, so we need to guarantee that a
slot is not overwritten.

For this case we introduce a new method on ExecutionContext
`newScopeWith1` that guarantees that we do not overwrite an existing
slot value. The current implementation makes a full copy of the
context, like is always done in the interpreted runtime. This could be
improved later.