Store correct last stored transaction id in transaction log file header on rotation
Previously on transactional log file rotation we were using last committed
transaction id as the id of the last transaction that can be found in the previous log file.
And that is not always correct since there is a window where transaction is already
stored in transaction log file, but not yet published as committed.
In case if the wrong header will be stored we will not gonna be able to find transactions
that were stored in older files since we will look for them in the incorrect place.

This PR change that behavior by using current committing transaction id
instead of last committed transaction id.
It's safe to do so since during rotation:
- we flush previous tx log file so we guarantee the durability of tx that still registered as committing;
- the operation is done under tx file lock so its impossible to race with
any ongoing commit