Add rudementary volumetric check-point triggering.
Currently it only *starts* a check-point process as soon as it looks
like there are logs that could be pruned.

In the future it would be interesting to have it use statistics to try
to predict the frequency at which log prunings happen, and adjust the
check-point frequency to match.