Convert trait to abstract class: LatestRuntimeVariablePlannerCompat...
By doing that, all constructor arguments are passed explicitly to
the super constructor and we can be sure those are initialized at
construction time. All remaining abstract vals are changed to be used
only in defs, not in concrete vals anymore. This makes this class not
be a fragile base class anymore.