Release the storecopy mutex after the store copy
The store copy request handler just queues up requests
on the netty pipeline and the release of the mutex must
be scheduled as a response to the completion of the actual
handling. This is done by registering on the future supplied
by netty.

Also did a general clean up of the code here with a better flow,
cleaner paragraphs, refactoring of methods and a few comments.